+++ lib/grid/concrete_grid.hpp
@@ -35,7 +35,7 @@ namespace Bempp
 {
 
 // Forward declarations
-template<size_t codim> class Entity;
+template<int codim> class Entity;
 class GridView;
 
 /**
